apps/btc_resonance_strip/requirements.txt
PySide6==6.7.2
websocket-client==1.7.0
numpy==2.1.1
